Add methods GET and PUT /dragons/{id}

+++ src/main/java/ru/danmax/soa_lab2_first_service/dto/response/LocationResponseDto.java
@@ -11,9 +11,9 @@ import ru.danmax.soa_lab2_first_service.entities.Location;
 @NoArgsConstructor
 @AllArgsConstructor
 public class LocationResponseDto {
-    private float x;
-    private float y;
-    private float z;
+    private Integer x;
+    private Double y;
+    private Integer z;
     private String name;
 
     public static LocationResponseDto convertToDTO(Location location) {
